requirements.txt
requests>=2.31.0
pretty_traceback>=1.1.2
typer>=0.9.0
rich>=13.7.0
promptheus-client>=0.19.0
pydantic>=2.5.2
python-dotenv>=1.0.0
typing-extensions>=4.8.0
pygments>=2.16.1
terminal-image>=0.7.0
pygments-markdown-lexer>=0.1.0.dev39
jsonschema>=4.19.1
yaml>=0.2.5
threaded>=0.0.5
prompt_toolkit>=3.0.43
simplification>=0.7.1